chore: remove balance calls for core
