+++ packages/appkit/src/client/appkit-basic.ts
@@ -68,6 +68,22 @@ export class AppKit extends AppKitCore {
     }
   }
 
+  public override async syncIdentity(
+    _request: Pick<AdapterBlueprint.ConnectResult, 'address' | 'chainId'> & {
+      chainNamespace: ChainNamespace
+    }
+  ) {
+    return Promise.resolve()
+  }
+
+  public override async syncBalance(_params: {
+    address: string
+    chainId: string | number | undefined
+    chainNamespace: ChainNamespace
+  }) {
+    return Promise.resolve()
+  }
+
   protected override async injectModalUi() {
     if (!isInitialized && CoreHelperUtil.isClient()) {
       await import('@reown/appkit-scaffold-ui/basic')
